Scratch buffer iteration model
Define how the scratch buffer is iterated within sessions and how
session boundaries are defined. "New" and "Clear" features, which
iterate the scratch buffer (and in the former case, starts a new
session). "Ab initio" iteration that creates the scratch buffer and
saves it on disk the first time, when one doesn't already exist.