Fix load session
Now that sessions can have any major mode, and therefore, file
extension, we need to compute the session filename from the filename
prefix and load such a file if present. We want to preserve the UX of
the user selecting just the folder in order to restore the session,
since from their perspective that already identifies it.