Consult Emacs for major mode and extension mapping
Formerly, we were inferring a mapping between template file extensions
and appropriate major modes, as well as a mapping between major modes
and template files, using methods peculiar to this package.

We now allow Emacs itself to set the major mode where possible, and in
cases where we need to infer it, we consult Emacs's ground-truth
`auto-mode-alist` to determine the appropriate major mode for a session
buffer based on the template file extension.

To infer the appropriate template given a major mode, we once again use
`auto-mode-alist`, by simply searching the template folder for any files
that have an extension recognizable to the major mode.

Addresses #13.