Attempt to iterate even if Git doesn't detect changes in known files
This is to support the possibility of adding and then committing
changes in files that aren't being visited, and which may have been
added externally to Mindstream.

This also removes a warning on nonzero exit code from shell command,
since that's being hit more often now (e.g. attempting to commit empty
changes) and it's not an error.