Generalize sessions to full Git repos
Formerly we needed to explicitly enter `mindstream-session-mode` for a
buffer to be considered part of the active session. Since sessions
naturally are associated with entire repos rather than individual
files, eliminate `mindstream-session-mode` in favor of a registry of
active session paths, so that a session may be begun from any file at
a certain path (and apply to that entire path).

All Mindstream-related keybindings are now in the global minor mode,
and we can also start and end sessions in existing Git repos.