Implement new "enter session" behavior
This first attempts to visit the MRU anonymous session for the
inferred major mode template. If there isn't one already open, it
attempts to open _all_ active anonymous sessions for the
template (visiting an arbitrary one, since there is no way to
determine "recency" for closed sessions). If there aren't any, then it
creates a new session.

In particular, this now disregards the anonymous session buffer name
and instead uses just the filesystem information to determine whether
a session is anonymous or not. That allows it to visit active
anonymous sessions whether they are currently open or not, and will
support the ability to have persistent sessions. By opening _all_
anonymous sessions for the inferred template, it also supports the
(as yet unimplemented) ability to have more than one active anonymous
session for a template at any time.