Update docs to account for session persistence and uniqueness
This just brings the docs up to date so that they remain accurate, but
it's likely that a more thorough reorganization would be beneficial.
As there are already more big changes planned (especially around
redefining "sessions" to correspond to Git branches for maximum
portability of Mindstream functionality), we will leave the more
thorough documentation update until that point.