Don't rely on session registry for anonymous sessions
Maintaining global state to identify active Mindstream sessions is
bug-prone, and we hope to eliminate it across the board in the future,
but for now we can at least eliminate it for anonymous sessions
specifically, as we always know their base path and can assume that
all sessions at this path are to be treated as active (i.e. versioned)
Mindstream sessions.

Fixes #28