Use `basic-save-buffer` instead of `save-buffer` as trigger
`basic-save-buffer` seems to be a more primitive way to save the
buffer that's called by `save-buffer` and perhaps by other actions
in Emacs that result in the buffer implicitly being saved. It may even
be called directly by other extensions (e.g. `super-save`). Using this
as a trigger for versioning should capture more of these cases and not
be limited to just user invocations of `C-x C-s`.

Fixes #29.