Use `:after` style advice for implicitly iterating
We don't need to do anything fancy here (e.g. modifying the arguments
or using the return value of the original function). We just need to
iterate every time a trigger function (e.g. `basic-save-buffer`) is
called, so `:after` style advice is sufficient. This form of advice is
likely also less intrusive in terms of how it appears in the stack
trace for the original function invocation.