File sessions under their start date
Instead of introducing date-based filing only in the archive, we now
file even anonymous sessions under their start date, so that having
the date in the name of the session is no longer necessary, and
archiving doesn't introduce a different directory structure than the
one at the anonymous path.

The change from the behavior introduced by date-based filing in the
archive folder is that the dates even there are now the start date,
i.e. the same as in the anonymous path, and not the _date of
archiving_. While the latter is useful to know, it is also not as
deterministic as the former. As far as the folder structure goes, we
have to pick one, and having them both be the same is simpler.