On second thought, don't open persistent sessions on startup
Since we can pull them up on demand anyway via "enter
session" (context-sensitive) and `mindstream-open` (context-free,
interactive), we can just make the latter more prominent via a
keybinding in the minor mode.

Starting all sessions at startup could potentially be a bit intrusive
as there may be sessions that contain multiple files, and without some
additional handling of this case, it would require interaction with
the user to determine which file to open in each session -- not
something we generally want to deal with on startup. Doing it
on-demand seems cleaner.