Show contents of template folders while loading a session
Recently, we made the change to save sessions under their template
name in the save session path. But we didn't modify loading of
sessions, so that this still lists the top-level contents of the save
session path, which currently are just template names.

The present commit fixes this so that loading lists the contents of
every subfolder of the save session path and then shows those full
paths in the completion menu. This means that the menu no longer shows
just session names but instead full session paths, and also, it
expects that sessions to be loaded must follow the standard
organization scheme.

This doesn't feel the most natural, but it at least fixes the
immediate bug. Issue #35 mentions a potentially better option to
explore.